new application version used the last api version

+++ TrentinoLocal_Android/servicelib/src/main/java/com/xpeppers/servicelib/bean/Coupon.java
@@ -7,14 +7,14 @@ import java.io.Serializable;
  * @since 2015/09/17
  */
 public class Coupon implements Serializable {
-    private double id;
+    private long id;
     private String code;
 
-    public double getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(double id) {
+    public void setId(long id) {
         this.id = id;
     }
